requirements.txt
streamlit>=1.36.0
openai>=1.30.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.30
pandas>=2.2.2
pydeck>=0.9.1
plotly>=5.22.0
streamlit-authenticator>=0.3.2
loguru>=0.7.2 